FastPCAModule combined with PSFSubtractionModule in PSFSubtractionPCA.py
